+++ spec/fixtures/01_trivial_assertion/in.rb
@@ -2,6 +2,7 @@ require 'test_helper'
 
 class BananaTest < ActiveSupport::TestCase
   test "is delicious" do
+    assert Banana
     assert Banana.new.delicious?
   end
 end