+++ spec/fixtures/03_trivial_refutation/in.rb
@@ -2,6 +2,7 @@ require 'test_helper'
 
 class KiwiTest < ActiveSupport::TestCase
   test "is not delicious" do
+    refute Kiwi
     refute Kiwi.new.delicious?
   end
 end