+++ spec/fixtures/03_trivial_refutation/out.rb
@@ -1,4 +1,7 @@
 require "spec_helper"
 RSpec.describe Kiwi do
-  it "is not delicious" { expect(Kiwi.new.delicious?).to be_falsey }
+  it "is not delicious" do
+    expect(Kiwi).to be_falsey
+    expect(Kiwi.new.delicious?).to eq(false)
+  end
 end